requirement.txt
absl-py==0.9.0
astor==0.8.1
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
bleach==1.5.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-auth==1.10.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.28.1
grpcio-tools==1.28.1
ipython==7.13.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyter-server==0.1.1
jupyterlab==1.2.5
jupyterlab-pygments==0.1.0
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
line-profiler==3.0.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.2
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
ordered-set==3.1.1
paho-mqtt==1.5.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.2.0
pyrsistent==0.15.7
python-dateutil==2.7.2
pytz==2018.4
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
urllib3==1.25.8
voila==0.1.20
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.0.1
